requirements.txt
pymongo==4.6.0
requests==2.31.0
flask==2.3.3
python-dotenv==1.0.0
